Refactored where some functions are defined. Now prompts the user to train the model if it can't be loaded in predict.py
